update file 29/10/2020

+++ database/migrations/2020_10_23_085622_create_san_pham_table.php
@@ -17,8 +17,8 @@ class CreateSanPhamTable extends Migration
             $table->increments('id');
             $table->string('ten_san_pham')->unique();
             $table->integer('so_luong')->nullable()->default(0);
-            $table->integer('gia_goc')->nullable()->default(0);
-            $table->integer('gia_sale')->nullable()->default(0);
+            $table->double('gia_goc')->nullable()->default(0);
+            $table->double('gia_sale')->nullable()->default(0);
             $table->text('mo_ta')->nullable();
             $table->string('hinh_anh')->nullable();
             $table->date('ngay_dang')->nullable();
